Create orca.py: extract OrcaSlicer logic into engine module (#324)
Extract all OrcaSlicer-specific code from slicer.py, profiles.py, and
init.py into a new orca.py engine module per ADR-006.

slicer.py becomes a thin dispatcher that routes to orca_slice_plate()
for OrcaSlicer or the existing cura path. Shared Docker utilities
(_ensure_docker_image, find_slicer, etc.) remain in slicer.py.

Backward-compatible re-exports ensure existing imports continue to work.
All 646 tests pass with zero lint/type errors.
