+++ changes/+create-orca-module.misc
@@ -1,0 +1,1 @@
+Extract OrcaSlicer-specific logic from ``slicer.py``, ``profiles.py``, and ``init.py`` into new ``orca.py`` engine module per ADR-006.
